Cleanup asset queuing
Assets are registered in the init hook and then enqueued in the respective hooks they should be fired in.  All asset code related to the block editor is compiled and placed in the build folder as a single javascript file. Blocks whose editor assets reside in this file can enqueue the registered handle.

Cleaned up enqueue file.
 - docblocks
 - camel casing

Distinguished between block-editor-assets and block-assets.